+++ src/dom/element_type.rs
@@ -33,7 +33,8 @@ pub trait ElementType {
     /// use svgdom::{Document, ElementType};
     ///
     /// let doc = Document::from_str("<svg><linearGradient/></svg>").unwrap();
-    /// let mut iter = doc.descendants().svg();
+    /// let mut iter = doc.descendants();
+    /// assert_eq!(iter.next().unwrap().is_referenced(), false); // root
     /// assert_eq!(iter.next().unwrap().is_referenced(), false); // svg
     /// assert_eq!(iter.next().unwrap().is_referenced(), true); // linearGradient
     /// ```